yoshino: tone: loire: locally cache the value of auth_id
The cmd used to receive auth_id from TZ in fpc_load_db_id can cause device crashes under certain unclear circumstances.
The auth_id value remains the same, unless one of the following happens:
1. A new fingerprint is enrolled
2. All enrolled fingerprints are deleted
So in order to reduce the number of times the cmd is sent, cache the value returned by fpc_load_db_id once the call is made.
For subsequent calls, simply rely on the cached value.
In case an enrollment happens, or a fingerprint is deleted, remove the cached value in order for the next call to repopulate the cache.
This should be sufficient to fix the random rebooting happening on yoshino, loire and tone devices.

Change-Id: I7815eafb0ef69b82dafe3581a7b103ec79aeb473