Properly reinitialize the TZ app in case of auth flood
Particularly on Yoshino (but may happen on others), when the
fingerprint related TZ app gets too many authentication failures,
it generates an unrecoverable error and disallows any further
authentication trial.
The only way to recover this condition is to completely stop and
reinitialize the TZ application.

To achieve this, split the BiometricsFingerprint::setActiveGroup
function into a subfunction __setActiveGroup, containing most of
its logic, that will be called in case the error condition above
described will happen.

This condition has never shown in case of process_enroll, hence
the handling in that case is ignored.