Ensure worker thread is free before sending cmds from "service" thread.
This introduced issues with navigation gestures: the TZ+phone would
crash often whenever a command is sent while it's in "blocking on
gestures" state.
Thus, before doing anything, the TZ must be guaranteed to not be in
polling state (by waking up the poll, and making sure
FPC_NAVIGATION_EXIT is called).

Why like this?
Instead of engineering a separate state for navigation gesture handling,
which involves switching to it at the right time (eg. when exiting
auth/enroll state) while being resilient against race conditions.
Those happen when the Service using this HAL calls another state
just as we are setState'ing to the navigation state. Mutexes, peeking
into the eventfd and the like are just not as elegant.

The upside of having this in a non-zero state (much like STATE_CANCEL)
means that any poll on the eventfd returns immediately; when an operation
is blocking on the irq and eventfd it'll wake up, finalize and give
control back to the worker_thread.

Signed-off-by: Marijn Suijten <marijns95@gmail.com>