BiometricsFingerprint: Presumably fixed potential deadlock.
At some point the HAL (and consequently entire system server) locked up
in a futex call.

Assumption:
While the service using this HAL should generally not call multiple
functions at the same time, it is possible that this happened and caused
a race condtion. Previously, setState was not protected by a mutex,
meaning that another thread could overwrite the eventfd *while* another
thread is waiting for the thread to transition to the desired state in
waitForState.
In this case it could well be a cancel() request that happens while
another thread is inside waitForState (for authenticate(), for example).
Here, something can be said for the resilience of the service itself: if
a HAL decides to do work on the main thread in one of the API functions
(instead of properly offloading it and returning ASAP) it should still
be able to cancel this operation.

For this case, protect the yet-unprotected setState with the same mutex,
and make sure it's already locked in nested scenarios.
This means, whenever a waitForState is happening, nothing can interrupt
it until the thread switches to the desired state (or dies).

Signed-off-by: Marijn Suijten <marijns95@gmail.com>