SAFETY: Use timeout in waitForState to prevent full system lockups.
Blocking HAL functions for too long takes the entire system down, which
is not what we want to do under _any_ circumstance.
Even though the HAL is supposedly free of race condtions and lockups, we
want to prevent against any coding mistake.
This is especially true around the fingerprint settings menu, where
locking the screen and authenticating cause a ton of state changes.
Crashing the HAL makes the entire problem very obvious to the user and
in the logs, while init takes care of restarting it in the end.

In the end it's all about making the HAL enter navigation state
consistently, aside from doing other operations. The system can call any
operation at any moment, for which gesture blocking must temporarily be
disabled to prevent issues in the TZ.

Signed-off-by: Marijn Suijten <marijns95@gmail.com>