Egistec: Ganges: Implement swipe gestures.
Egistec is much simpler than FPC in that it doesn't crash when calling
setup commands while the TZ app is in navigation mode.
All it needs to do is call the GetNavEvent command followed by polling
the device. Whenever the IRQ is raised in the kernel the same function
can be called again to retrieve an integer with a keycode. This keycode
uses the same values as the ones reported through uinput, but the
directions are wrong and need a local conversion.

Signed-off-by: Marijn Suijten <marijns95@gmail.com>