BiometricsFingerprint: Simplify state transitions with mutexes.
This change implements the new state transition mechanism explained in
the PR description here:
https://github.com/sonyxperiadev/vendor-sony-oss-fingerprint/pull/50

The previous system was unable to work correctly with an eventfd that
cannot be read without clearing the value, nor safely written without
incrementing the previous one.

For that reason the desired next state is kept in a separate variable,
and any access is protected by a mutex. When a new event is available
the eventfd is set to a non-zero value to wake up any pollers. When the
worker thread gets to read the next state the eventfd is cleared and the
condition variable notified, in case the requester needs to wait for the
thread to have finished the current operation.

Any "implicit hacks" with non-zero states written to the eventfd are now
removed since the eventfd is filled with a non-zero state regardless of
what is requested.
The cancel state becomes obsolete and is removed as well, since the
desired state can now be requested directly.

Signed-off-by: Marijn Suijten <marijns95@gmail.com>