BiometricsFingerprint: Extract thread-handling to separate class.
The HAL implementation is already getting crowded, and this
functionality may well be used elsewhere. For that reason it is
extracted to a separate class and file, just like the WorkerThread for
Egistec implementations.

The implementations are not unified because Egistec benefits nothing
from the extra synchronization. It operates perfectly fine when setup
commands are called while gesture-navigation polling is ongoing,
contrary to FPC which crashes _the entire device_.
Besides, the legacy C FPC implementations still require access to the
waitForEvent method and fds, which is a pain to share properly and clean
up.
If Egistec turns out to need this sync as well, it can simply be
replaced. waitForEvent and eventfd/epoll setup can be added on top in a
subclass.

Signed-off-by: Marijn Suijten <marijns95@gmail.com>