BiometricsFingerprint: CRASHFIX: Always call update_template.
Authentication crashed consistently on tama the next time after
presenting an unknown fingerprint. While I have been able to address the
recovery by introducing a sleep between closing and reopening
everything, it is of course much better to address the cause directly.

As it turns out, the TZ really wants update_template to be called (and
perhaps store_user_db) even when the identified print is zero (meaning
no match found). This is fixed by always calling these functions outside
the if (print_id > 0) block.
Besides, the function is updated to represent a new capability that
tells whether the DB has been changed and needs storing. This field
always seems to contain 1 anyway, but is added for completeness
regardless.

Signed-off-by: Marijn Suijten <marijns95@gmail.com>