yoshino: Disable fingerprint gestures.
Yoshino devices crash after a minute with this feature enabled. There is
no way for us to debug or fix it, as it is most likely a TZ-app issue
which cannot be resolved. Disable the feature completely to prevent
calling any of these navigation functions giving the user an unusable
device.

Signed-off-by: Marijn Suijten <marijns95@gmail.com>