BiometricsFingerprint: uinput: fall back to write on older kernels
On sailfish XA2 we are currently on 4.4 kernel and this kernel
does not support UI_DEV_SETUP. We need to use the newer fingerprint
HAL since flashing Android 9 with Emma and then SailfishOS currently
breaks fingerprint on SailfishOS. Backporting the fixes for that would
also be an option but staying with upstream vendor-sony-oss-fingerprint
seems better.

So fallback to write if ioctl UI_DEV_SETUP fails in order to avoid
a crashing android.hardware.biometrics.fingerprint service.