Android.mk: Add support for Kumano devices.
Add the usual definitions and make sure the "current" Egistec HAL is
hosted in service.cpp.

Unfortunately the name of the TZapp changed from egisap32 to egista, and
the files moved to /odm/firmware. For simplicity a preprocessor macro is
used to switch these variables. If actual commands change in the future
a more sophisticated method should be introduced with separate types,
likely switching between these in service.cpp and passing it as
argument.

Signed-off-by: Marijn Suijten <marijns95@gmail.com>