egistec/current: Retrieve and print hardware ID.
This is a new function available on 6xx and will be used to switch logic
based on the hardware id, instead of preprocessor definitions.
(effectively making a single binary compatible across different devices)

This function returns an ID of 0 on unsupported devices, such as the
et512 series found in Ganges.

Signed-off-by: Marijn Suijten <marijns95@gmail.com>