egistec: current: Disable power control on Kumano
Kumano devices are the first with in-kernel (copyleft) power management
support. Unfortuately this only streches to HAL starts, the device is
left on after that indefinitely. Toggling power to save some microAmps
while running - which works on all platforms till now - make the sensor
on this device completely unresponsive, request HW resets and the like.
This is an unworkable solution that has to be disabled.
The preprocessor flag is spread to all platforms including non-egistec
ones in case we want to reuse it in the future for FPC.

Aside from that Kumano is the first platform where power management
remotely makes any sort of sense since there is no gesture support and
the sensor could realistically be off when the device is in use.

Some autoformats slipped in, too.

Signed-off-by: Marijn Suijten <marijns95@gmail.com>