egistec: current: HAL: Implement retry image acquire mechanism
This fingerprint sensor seems to be very picky about its own
electrical conditions and even a small electromagnetic charge
seems to be messing the imager state.

For this reason, retry getting the fingerprint image for up to
7 times after resetting the sensor: this makes sure that the
supply to it is 99.9% stable and gives it the best chance to
acquire a full *good* image. Of course, if the sensor is dirty
for real, it will keep returning a ImagerDirty event.

Also, during fingerprint enrollment, keep calling the imager
calibration function in the TZ driver, so that we tend to
increase the quality of the acquired reference image as much
as possible: in this case, the subsequent authentication will
be less picky, since the initial reference quality is higher.

While at it, also remove the DirtOnSensor result handling during
fingerprint enrollment: this tends to lockup the sensor for up
to 10s econds.