BiometricsFingerprint: Drop unused <chrono> header
This header was initially (and erroneously) introduced by me in 0c91cfb,
and later forgotten to be moved into the new SynchronizedWorkerThread
file in 6cc58d1, also by me.

It is not necessary to explicitly include this header because
std::condition_variable::wait_for() takes an std::chrono::duration; that
declaration (in <condition_variable>) can only be made if the <chrono>
header is included prior.

Signed-off-by: Marijn Suijten <marijns95@gmail.com>