BiometricsFingerprint: Correct uses of ErrorFilter, always call .Resume
This PR cleans up misuses of ErrorFilter(-1) which is a case that
doesn't exist:

    ALOGE("An unknown error returned from fingerprint vendor library:
    %d", error);

This is completely misleading, especially when it regards an EINVAL.
Replace these cases with an appropriate RequestStatus::.

Furthermore check errors more rigorously on functions that return them;
these are the exact candidates that ErrorFilter exists for, so better
use it.
Finally fix one case of such an undesired short-circuit in ::enumerate,
where an error in fpc_get_print_index() would not cause the loop to
return to idle state, hence breaking swipe gestures temporarily until
the next event.

Signed-off-by: Marijn Suijten <marijns95@gmail.com>