BiometricsFingerprint: Delete all fingerprints when fid=0
For some odd reason android ceased to pass fid=0 when deleting all
fingerprints - which I recall happening before when working on Egistec -
likely as an indication to _really_ clean up in case Androids accounting
of fingers (in /data/system/users/$gid/settings_fingerprint.xml) gets
desynced from the proprietary fingerprint database.

This special value is part of the HIDL specification and was found to be
an issue on Sailfish OS, which properly utilizes this functionality.

Implement a loop over all known prints in much the same way as
Egistec/current, and test it by manually setting fid=0; to see all
prints disappear.

Signed-off-by: Marijn Suijten <marijns95@gmail.com>