FPC: Remove unused (always zero) database length (Kitakami leftover)
This is apparently internal to Kitakami, that's not supported by this
HAL at all for _years_. In the event that one ever wishes to do so, it
should be retrieved internally in fpc_store_user_db as the FIXME comment
explains.

Remove the call now, as it only leads to unnecessary confusion while
reading the logs and code.