Clean up typedefs when tagged `struct <structname>` is used
It is typical to either `typdef` a struct - in which case the struct
does not need a name itself - or use the `struct` tag in front of the
struct name.  However, this codebase was mixing and matching both,
giving the struct name the `_t` suffix as well as then using it either
with or without `struct` tag.  Clean this up by at least removing the
`typedef` and `_t` suffix from all structs that are mostly or
exclusively used as tagged structs.